Worked on spint2 and sprint3

+++ src/app/shared/base-layout/base-layout.component.ts
@@ -1,17 +1,28 @@
 import { Router } from "@angular/router";
 import { CookieService } from "ngx-cookie-service";
 import { Component, OnInit } from "@angular/core";
+import {
+	ConfirmationService,
+	ConfirmEventType,
+	MessageService,
+} from "primeng/api";
 
 @Component({
 	selector: "app-base-layout",
 	templateUrl: "./base-layout.component.html",
 	styleUrls: ["./base-layout.component.css"],
+	providers: [ConfirmationService, MessageService],
 })
 export class BaseLayoutComponent implements OnInit {
 	sessionName: string;
 	year: number;
 
-	constructor(private cookieService: CookieService, private router: Router) {
+	constructor(
+		private cookieService: CookieService,
+		private router: Router,
+		private messageService: MessageService,
+		private confirmationService: ConfirmationService
+	) {
 		this.sessionName = this.cookieService.get("session_name");
 		this.year = Date.now();
 	}
@@ -19,7 +30,32 @@ export class BaseLayoutComponent implements OnInit {
 	ngOnInit(): void {}
 
 	logout() {
-		this.cookieService.deleteAll();
-		this.router.navigate(["/session/login"]);
+		this.confirmationService.confirm({
+			message: "Are you sure that you want to proceed?",
+			header: "Log out Confirmation",
+			icon: "pi pi-exclamation-triangle",
+			accept: () => {
+				this.cookieService.deleteAll();
+				this.router.navigate(["/session/login"]);
+			},
+			reject: (type: any) => {
+				switch (type) {
+					case ConfirmEventType.REJECT:
+						this.messageService.add({
+							severity: "info",
+							summary: "Cancelled",
+							detail: "Logout cancelled",
+						});
+						break;
+					case ConfirmEventType.CANCEL:
+						this.messageService.add({
+							severity: "info",
+							summary: "Cancelled",
+							detail: "Logout cancelled",
+						});
+						break;
+				}
+			},
+		});
 	}
 }
